requirements/base.txt
appdirs==1.4.3
configparser==3.5.0
Django==1.10.6
django-rest-swagger==0.3.10
djangorestframework==3.6.2
enum34==1.1.6
flake8==3.3.0
mccabe==0.6.1
packaging==16.8
pycodestyle==2.3.1
pyflakes==1.5.0
pyparsing==2.2.0
PyYAML==3.12
requests==2.13.0
six==1.10.0
whitenoise==3.3